* server/bitmap_info.h: drop OGL-specific members.         * libbase/image.{cpp,h}: cleanup the image_base interface:           - make more members private to reduce chaces of buffer size             and associated size-describing members mismatch (still more             to do)           - add more methods from statics which do not have access             to now-private members         * backend/render_handler_ogl.cpp: add OGL-specific members           to bitmap_info_ogl, update uses of image_base.         * backend/render_handler_agg.cpp, libbase/embedVideoDecoderFfmpeg.cpp,           libbase/embedVideoDecoderGst.cpp, libbase/image_filters.cpp,           server/FreetypeGlyphsProvider.cpp, fill_style.cpp,           server/fontlib.{cpp,h}, server/parser/BitmapMovieDefinition.cpp,           server/swf/tag_loaders.cpp, server/asobj/NetStream.cpp,           server/asobj/NetStreamFfmpeg.cpp, backend/render_handler_cairo.cpp:           Update uses of image_base

+++ server/fontlib.h
@@ -88,7 +88,7 @@ namespace fontlib {
 	// Applies to fonts processed *after* this call only.
 	// The "nominal" size is perhaps around twice the
 	// average glyph height.
-	void	set_nominal_glyph_pixel_size(int pixel_size);
+	void	set_nominal_glyph_pixel_size(size_t pixel_size);
 
 	/// Clean up the font library
 	void	clear();
